requirements.txt
Django==2.2.6
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.10.3
mongoengine==0.18.2
pymongo==3.9.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
